Construct isolated namespaces for schemas

+++ discovery/api/handlers/registry.py
@@ -10,74 +10,29 @@ from tornado.escape import json_decode
 
 from discovery.api.es.doc import SchemaClass, Schema
 
-from .base import APIBaseHandler
-
-
-def github_authenticated(func):
-    '''
-        RegistryHandler Decorator
-    '''
-
-    def _(self, *args, **kwargs):
-
-        if not self.current_user:
-
-            self.send_error(
-                message='login with github first',
-                status_code=401
-            )
-            return
-
-        func(self, *args, **kwargs)
-
-    return _
-
-
-def permisson_verifeid(func):
-    '''
-        RegistryHandler Decorator
-
-        - Validate user permission to edit a schema
-        - Only the owner has the permission
-
-    '''
-
-    def _(self, prefix, *args, **kwargs):
-
-        schema = Schema.get(id=prefix, ignore=404)
-
-        if not schema:
-            self.send_error(404)
-            return
-
-        if schema['_meta'].username != self.current_user:
-            self.send_error(403)
-            return
-
-        func(self, prefix, *args, **kwargs)
-
-    return _
+from .base import APIBaseHandler, github_authenticated
 
 
 class RegistryHandler(APIBaseHandler):
     '''
         Registered Schema Repository
 
-        Check  - HEAD ./api/registry/<schema>
+        Check  - HEAD ./api/registry/<namespace>
         Create - POST ./api/registry
         Fetch  - GET ./api/registry
         Fetch  - GET ./api/registry?user=<username>
-        Fetch  - GET ./api/registry/<schema>/<class>
-        Remove - DELETE ./api/registry/<schema>
+        Fetch  - GET ./api/registry/<namespace>/<curie>
+        Remove - DELETE ./api/registry/<namespace>
 
     '''
 
-    def head(self, prefix):
+    def head(self, namespace):
         '''
-            Check the existance of a schema by its prefix.
+            Check the existance of a schema by its namespace.
         '''
-
-        if prefix == 'schema' or Schema.get(id=prefix, ignore=404):
+        if namespace == 'schema':
+            self.set_status(200)
+        elif Schema.get(id=namespace, ignore=404):
             self.set_status(200)
         else:
             self.set_status(404)
@@ -87,17 +42,16 @@ class RegistryHandler(APIBaseHandler):
         '''
             Create a new schema entry in the registry.
         '''
-
         args = json_decode(self.request.body)
-        prefix = args['namespace']
+        namespace = args['namespace']
         url = args['url']
 
-        assert prefix != 'schema', "cannot rewrite core schema."
+        assert namespace != 'schema', "cannot rewrite core schema."
 
-        if Schema.get(id=prefix, ignore=404):
+        if Schema.get(id=namespace, ignore=404):
 
             self.send_error(
-                reason=f"'{prefix}'' is already registered.",
+                reason=f"'{namespace}'' is already registered.",
                 status_code=403
             )
             return
@@ -105,41 +59,41 @@ class RegistryHandler(APIBaseHandler):
         schema_doc = requests.get(url, timeout=5)
         schema_doc.raise_for_status()
         schema_parser = self.get_parser(schema_doc.json())
-        schema_classes = SchemaClass.import_from_parser(schema_parser)
+        schema_classes = SchemaClass.import_classes(schema_parser, namespace)
 
         for klass in schema_classes:
             klass.save()
 
         schema = Schema(**{
-            "meta": {"id": prefix},
-            "context": schema_parser.context.get(prefix, None),
+            "meta": {"id": namespace},
+            "context": schema_parser.context,
         })
         schema._meta.url = url
         schema._meta.username = self.current_user
         schema.encode_raw(schema_doc.text)
-        schema.save()
 
         logger = logging.getLogger(__name__)
-        logger.info("Saved '%s'.", prefix)
+        logger.info("Saved schema namespace '%s'.", namespace)
 
-        self.set_status(201)
+        self.set_status(200)
         self.finish({
             'success': True,
+            'result': schema.save(),
             'total': len(schema_classes),
             'url': self.request.full_url() + '/' + schema.meta.id,
         })
 
-    def get(self, prefix=None, label=None):
+    def get(self, namespace=None, curie=None):
         '''
             Access the registry.
 
             - List all schemas.
             - List schemas by a user.
-            - List a schema by its prefix.
-            - List a class by its name and prefix.
+            - List a schema by namespace.
+            - List a class by its namespace and curie.
         '''
 
-        if not prefix:
+        if namespace is None:
 
             user = self.get_query_argument('user', None)
 
@@ -151,14 +105,15 @@ class RegistryHandler(APIBaseHandler):
             else:
                 search = search.query("match_all")
 
+            schema_context = dict()
+            for schema in search.scan():
+                schema_context.update(schema.context.to_dict())
+
             self.write({
                 "total": search.count(),
-                "context": {
-                    schema.meta.id: schema.context
-                    for schema in search.scan()
-                },
+                "context": schema_context,
                 "hits": [{
-                    "prefix": schema.meta.id,
+                    "namespace": schema.meta.id,
                     "url": schema['_meta'].url,
                 } for schema in search.scan()]
             })
@@ -166,22 +121,19 @@ class RegistryHandler(APIBaseHandler):
 
         result = {}
 
-        if prefix != 'schema':
+        if namespace != 'schema':
 
-            schema = Schema.get(id=prefix, ignore=404)
-
+            schema = Schema.get(id=namespace, ignore=404)
             if not schema:
                 self.send_error(404)
                 return
-
             result['url'] = schema['_meta'].url
 
-        if not label:
+        if curie is None:
 
-            search = SchemaClass.search().query("match", prefix=prefix)
+            search = SchemaClass.search().query("match", namespace=namespace)
             search.params(rest_total_hits_as_int=True)
 
-            result['name'] = prefix
             result['total'] = search.count()
             result['context'] = Schema.gather_contexts()
             result['hits'] = [klass.to_dict()
@@ -190,7 +142,7 @@ class RegistryHandler(APIBaseHandler):
             self.write(result)
             return
 
-        klass = SchemaClass.get(id=f"{prefix}:{label}", ignore=404)
+        klass = SchemaClass.get(id=f"{namespace}::{curie}", ignore=404)
 
         if not klass:
             self.send_error(404)
@@ -200,15 +152,23 @@ class RegistryHandler(APIBaseHandler):
         return
 
     @github_authenticated
-    @permisson_verifeid
-    def delete(self, prefix):
+    def delete(self, namespace):
         '''
         Delete a schema and its classes by its prefix.
         '''
+        schema = Schema.get(id=namespace, ignore=404)
 
-        sch = Schema.get(id=prefix)
+        if not schema:
+            self.send_error(404)
+            return
+
+        if schema['_meta'].username != self.current_user:
+            self.send_error(403)
+            return
+
+        sch = Schema.get(id=namespace)
         sch.delete()
 
-        SchemaClass.delete_by_schema(prefix)
+        SchemaClass.delete_by_schema(namespace)
 
         Index('discover_class').refresh()
